implementation patch

+++ src/Pages/Bookings/Booking.jsx
@@ -2,6 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../Components/AuthProvider/AuthProvider";
 import {FaRegTimesCircle} from 'react-icons/fa'
 import PageHeader from "../../Components/PageHeader/PageHeader";
+import Swal from "sweetalert2";
 
 const Booking = () => {
   const { users } = useContext(AuthContext);
@@ -18,17 +19,69 @@ const Booking = () => {
 
 
   const handleDelete=(id)=>{
-        fetch(`http://localhost:5000/bookings/${id}`,{
-            method:'DELETE'
-        })
-        .then((res) => res.json())
-        .then(data=>{
-            if(data.deletedCount>0){
-              const remaning=bookings.filter(book=>book._id!==id)
-             
-              setBookings(remaning)
+
+
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+            fetch(`http://localhost:5000/bookings/${id}`,{
+                method:'DELETE'
+            })
+            .then((res) => res.json())
+            .then(data=>{
+                if(data.deletedCount>0){
+                  const remaining=bookings.filter(book=>book._id!==id)
+                  
+                  setBookings(remaining)
+                }
+            })
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+        }
+      })
+
+        
+  }
+  const handleConfrim=(id)=>{
+
+      fetch(`http://localhost:5000/bookings/${id}`,
+      {
+        method:'PATCH',
+        headers:{
+            'content-type': 'application/json'
+        },
+        body: JSON.stringify({status:'confirm'})
+      }
+      )
+      .then(res=>res.json())
+      .then(data=>
+        {
+            if(data.modifiedCount>0){
+                const remaining=bookings.filter(booking=>booking._id!==id);
+                const updated=bookings.find(booking=>booking._id===id)
+                updated.status='confirm'
+                const result=[updated,...remaining]
+                Swal.fire({
+                    position: 'top-end',
+                    icon: 'success',
+                    title: 'Confirmed Successfully',
+                    showConfirmButton: false,
+                    timer: 1500
+                  })
+                setBookings(result)
             }
-        })
+        }
+        )
   }
 
   return (
@@ -66,9 +119,13 @@ const Booking = () => {
                  </td>
 
                  <td className="border-b border-gray-200 bg-white px-5 py-5 text-sm">
-                   <span className="rounded-full bg-yellow-200 px-3 py-1 text-xs font-semibold text-yellow-900">
-                     Suspended
-                   </span>
+                 {
+                    bookCard.status==='confirm'?   <button  className="rounded-full bg-green-200 px-3 py-1 text-xs font-semibold text-yellow-900">
+                    Confirm
+                  </button>:  <button onClick={()=>handleConfrim(bookCard._id)} className="rounded-full bg-yellow-200 px-3 py-1 text-xs font-semibold text-yellow-900">
+                     Pending
+                   </button>
+                 }
                  </td>
                </tr>
                 )}
